Add synsets for 'COVID-19', 'MERS', 'coronavirus disease'
This PR adds synsets for COVID-19 and terminology variants as on
Wikipedia. It also adds a new super-concept 'coronavirus disease' which
includes the existing 'SARS' synset and a synset for 'MERS'.

While this PR slightly contradicts the rule of avoiding synsets that are
already in Wikipedia, in this case it may be important enough to break
the rule.