Fix some capitalization
When searching for proper nouns, I found a number of terms that are
capitalized but should not generally be. The main kinds of changes are

1. **Binomial names**: Binomial names should capitalize the first word
and not the second, e.g., 'Anemone Canadensis' => 'Anemone canadensis'
2. **Role words**: Role words are always common nouns and should be
lowercased, e.g., 'president of the United States'. We apply this even
to formal titles as the synsets refer to the person holding the role,
e.g., 'the chancellor of the exchequer' holds the title of 'Chancellor
of the Exchequer'
3. **Brand/place names used generically**: For example 'armageddon' to
refer to 'any catastrophically destructive battle' but still
'Armageddon' for '(New Testament) the scene of the final battle between
the kings of the Earth at the end of the world'