Add links for species to Wikidata
This is linking is mostly based on the taxonomic names, which was then
further reviewed. Of these mappings I confirmed 4335 automatically by
using the existing genus links (i.e., 'Panthera leo' was accepted as the
parent 'Panthera' was already linked and the parent was the same). A
further 766 could not be confirmed automatically and were manually
confirmed. A second evaluation showed that 143 were not compatible with
the existing links, and this was evaluated and in nearly all cases this
was due to a secondary scientific name in the OEWN synset, with a small
number of errors in OEWN also corrected.